Fix ensemble tests with changes from tree PR
