+++ .python-version
@@ -1,0 +1,8 @@
+3.10.0
+3.9.5
+3.8.10
+3.7.10
+pypy3.7-7.3.9
+pypy2.7-7.3.4
+2.7.18
+3.11.0rc2
