+++ newrelic/config.py
@@ -2799,91 +2799,91 @@ def _process_module_builtin_defaults():
     _process_module_definition(
         "sklearn.tree.tree",
         "newrelic.hooks.mlmodel_sklearn",
-        "instrument_sklearn_models",
+        "instrument_sklearn_tree_models",
     )
 
     _process_module_definition(
         "sklearn.ensemble._bagging",
         "newrelic.hooks.mlmodel_sklearn",
-        "instrument_sklearn_models",
+        "instrument_sklearn_ensemble_models",
     )
 
     _process_module_definition(
         "sklearn.ensemble.bagging",
         "newrelic.hooks.mlmodel_sklearn",
-        "instrument_sklearn_models",
+        "instrument_sklearn_ensemble_models",
     )
 
     _process_module_definition(
         "sklearn.ensemble._forest",
         "newrelic.hooks.mlmodel_sklearn",
-        "instrument_sklearn_models",
+        "instrument_sklearn_ensemble_models",
     )
 
     _process_module_definition(
         "sklearn.ensemble.forest",
         "newrelic.hooks.mlmodel_sklearn",
-        "instrument_sklearn_models",
+        "instrument_sklearn_ensemble_models",
     )
 
     _process_module_definition(
         "sklearn.ensemble._iforest",
         "newrelic.hooks.mlmodel_sklearn",
-        "instrument_sklearn_models",
+        "instrument_sklearn_ensemble_models",
     )
 
     _process_module_definition(
         "sklearn.ensemble.iforest",
         "newrelic.hooks.mlmodel_sklearn",
-        "instrument_sklearn_models",
+        "instrument_sklearn_ensemble_models",
     )
 
     _process_module_definition(
         "sklearn.ensemble._weight_boosting",
         "newrelic.hooks.mlmodel_sklearn",
-        "instrument_sklearn_models",
+        "instrument_sklearn_ensemble_models",
     )
 
     _process_module_definition(
         "sklearn.ensemble.weight_boosting",
         "newrelic.hooks.mlmodel_sklearn",
-        "instrument_sklearn_models",
+        "instrument_sklearn_ensemble_models",
     )
 
     _process_module_definition(
         "sklearn.ensemble._gb",
         "newrelic.hooks.mlmodel_sklearn",
-        "instrument_sklearn_models",
+        "instrument_sklearn_ensemble_models",
     )
 
     _process_module_definition(
         "sklearn.ensemble.gradient_boosting",
         "newrelic.hooks.mlmodel_sklearn",
-        "instrument_sklearn_models",
+        "instrument_sklearn_ensemble_models",
     )
 
     _process_module_definition(
         "sklearn.ensemble._voting",
         "newrelic.hooks.mlmodel_sklearn",
-        "instrument_sklearn_models",
+        "instrument_sklearn_ensemble_models",
     )
 
     _process_module_definition(
         "sklearn.ensemble.voting_classifier",
         "newrelic.hooks.mlmodel_sklearn",
-        "instrument_sklearn_models",
+        "instrument_sklearn_ensemble_models",
     )
 
     _process_module_definition(
         "sklearn.ensemble._stacking",
         "newrelic.hooks.mlmodel_sklearn",
-        "instrument_sklearn_models",
+        "instrument_sklearn_ensemble_models",
     )
 
     _process_module_definition(
         "sklearn.ensemble._hist_gradient_boosting.gradient_boosting",
         "newrelic.hooks.mlmodel_sklearn",
-        "instrument_sklearn_models",
+        "instrument_sklearn_ensemble_models",
     )
 
     _process_module_definition(
