+++ newrelic/hooks/mlmodel_sklearn.py
@@ -47,6 +47,7 @@ def _wrap_method_trace(module, _class, method, name=None, group=None):
             # Set the _nr_wrapped attribute to denote that this method is no longer wrapped.
             setattr(trace, wrapped_attr_name, False)
 
+        # breakpoint()
         return return_val
 
     wrap_function_wrapper(module, "%s.%s" % (_class, method), _nr_wrapper_method)
@@ -56,6 +57,7 @@ def _nr_instrument_model(module, model_class):
     for method_name in METHODS_TO_WRAP:
         if hasattr(getattr(module, model_class), method_name):
             # Function/MLModel/Sklearn/Named/<class name>.<method name>
+            # breakpoint()
             name = "MLModel/Sklearn/Named/%s.%s" % (model_class, method_name)
             _wrap_method_trace(module, model_class, method_name, name=name)
 
@@ -63,6 +65,7 @@ def _nr_instrument_model(module, model_class):
 def _instrument_sklearn_models(module, model_classes):
     for model_cls in model_classes:
         if hasattr(module, model_cls):
+            # breakpoint()
             _nr_instrument_model(module, model_cls)
 
 
@@ -73,4 +76,28 @@ def instrument_sklearn_tree_models(module):
         "ExtraTreeClassifier",
         "ExtraTreeRegressor",
     )
+    _instrument_sklearn_models(module, model_classes)
+
+
+def instrument_sklearn_ensemble_models(module):
+    model_classes = (
+        "AdaBoostClassifier",
+        "AdaBoostRegressor",
+        "BaggingClassifier",
+        "BaggingRegressor",
+        "ExtraTreesClassifier",
+        "ExtraTreesRegressor",
+        "GradientBoostingClassifier",
+        "GradientBoostingRegressor",
+        "HistGradientBoostingClassifier",
+        "HistGradientBoostingRegressor",
+        "IsolationForest",
+        "RandomForestClassifier",
+        "RandomForestRegressor",
+        "StackingClassifier",
+        "StackingRegressor",
+        "VotingClassifier",
+        "VotingRegressor",
+    )
+    # breakpoint()
     _instrument_sklearn_models(module, model_classes)
